Moved querySelector checking, falls back on John Resig's Sizzle (http://github.com/jeresig/sizzle/tree/master) if document.querySelectorAll is not available.
removed the try/catch actions from respondToReadyState so that it won't silently error if the user of this library has an error in their callback functions

check if key is a function in Element creator before adding it as an attribute. this should ignore all of the Object.addMethods functions that anyone might do.


git-svn-id: svn+ssh://paularms.com/home/989/data/svn/jsdev/trunk@9 6bdbecb3-7c55-0410-a0e7-a0c93fd440fa